+++ packages/grid/src/features/standard-table/components/StandardTableCell.tsx
@@ -70,6 +70,7 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
     onChange,
     onKeyDown: onKeyDownCell,
     disableGridCell,
+    disableGridCellFocus,
     zIndex,
   } = useColumnConfigById(columnId);
 
@@ -190,7 +191,9 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
       }}
     >
       <StandardTableCellUi
-        enableGridCell={enableGridCell && !disableGridCell}
+        enableGridCell={
+          enableGridCell && !disableGridCell && !disableGridCellFocus
+        }
         gridCellRequiredProps={gridCell.requiredProps}
         isEditing={gridCell.isEditing}
         width={width}