+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 
-## Next release
+## Next
 
 ### StandardTable
+
 - `bannerError` is added to standard table and is prioritized before error property if given. Error is now displayed in a ResultListBanner.
 - `error and errorLabel` are now displayed in an ResultListBanner instead of ErrorScreen.
+- Column configs now have `onKeyDown` option.
+
+```
+onKeyDown: (event, { item, columnId }) => {
+  if (event.key === " ") {
+    setActive(item, !item.active);
+  }
+},
+```
+
 ## 14.0.2
 
 - `Drawer` prop `zIndex` has been added back.